New sigma_compiler! syntax
Instead of six lists of identifiers, there are now just two: a list of
identifiers for Scalars and a list of identifiers for Points.  Each
identifier can be preceded by zero or more tags (`pub`, `rand`, `cind`,
`const`, `vec`).

A `TaggedIndent` representing a `Scalar` can be preceded by:
 - (nothing)
 - `pub`
 - `rand`
 - `vec`
 - `pub vec`
 - `rand vec`

A `TaggedIndent` representing a `Point` can be preceded by:
 - (nothing)
 - `cind`
 - `const`
 - `cind const`
 - `vec`
 - `cind vec`
 - `const vec`
 - `cind const vec`