Refactor sigma_compiler_derive into sigma_compiler_derive and sigma_compiler_core
All of the work is done in the sigma_compiler_core crate, which is just
a normal library crate, not a proc-macro crate, so that it's more easily
developed and tested.

sigma_compiler_derive is then just a tiny wrapper.

h/t https://medium.com/data-science/nine-rules-for-creating-procedural-macros-in-rust-595aa476a7ff