Start making and using the StatementTree
The statements in the ZKP form a tree.  The leaves are basic
statements of various kinds; for example, equations or inequalities
about Scalars and Points.  The interior nodes are combiners: `And`,
`Or`, or `Thresh` (with a given constant threshold).  A leaf is true
if the basic statement it contains is true.  An `And` node is true
if all of its children are true.  An `Or` node is true if at least
one of its children is true.  A `Thresh` node (with threshold `k`) is
true if at least `k` of its children are true.