Be able to send extra generated public values along with the proof
We need to be able to send extra variables (not necessarily appearing in
the sigma_compiler macro invocation itself, since they may be generated
by the sigma_compiler itself) from the prover to the verifier along with
the proof.  These could include commitments to bits in range proofs, for
example.