Only emit code to check consistency of substitutions if we're in the root disjunction branch
For example, if the statement contains "y = x + 1" _not_ inside an "Or",
then both substitute "y" with "x+1" in the StatementTree, and also emit
code that returns VerificationFailure if the Witness contains x and y
that does not satisfy y = x + 1.

On the other hand, if the statement contains "y = x + 1" _inside_ an
"Or", then do the substitution, but do not emit the checking code, since
it may not be the case that that part of the StatementTree is intended
to be true.

This commit is not formatted so as to minimize the diff.  The next
commit will run "cargo fmt" with no code changes.