In a disjunction branch, convert statements about equality of public Scalars to statements about equality of public Points by multiplying each side by a generator
When we're not in a disjunction, we just emit code to directly check the
equality of the Scalars, as before.  But inside, the equality may not be
true.

So inside a disjunction, change statements like `b = 2*a - 3` into
`b*A = (2*a - 3)*A`, where `A` is one of the declared cind Points.

This commit is not formatted so as to minimize the diff.  The next
commit will run "cargo fmt" with no code changes.