Broaden the places where rand Scalars can appear
Before, a rand Scalar could only appear one time in total over
all of the statements in the ZKP.  If one appeared more than once,
it would not be considered "rand" for the purposes of recognizing
Pedersen commitments.

This was overly conservative, and the rule was indeed violated with
CMZ credentials, where the statements have the rand values appearing
(typically) twice each.

Now, rand Scalars can appear multiple times in linear combination
statements, but cannot appear at all (with a compile-time error) in
range or not-equals statements (where it never made sense for them to
appear anyway).